requirements/base.txt
Django==1.6.9
django-registration==1.0
django-forms-bootstrap==3.0.1
mongoengine==0.8.7
pymongo==2.7.1
mongodbforms==0.3
git+git://github.com/jschrewe/django-mongoadmin.git#egg=django-mongoadmin
git+git://github.com/snormore/django-mongoreversion.git#egg=django-mongoreversion
django-extra-views==0.6.4
python-dateutil==2.2
elasticsearch==1.2.0
elasticutils==0.9.1
pyelasticsearch==0.7.1
blinker==1.3
Pillow==2.5.3
django-ckeditor-updated==4.4.0
factory-boy==2.3.1
celery[mongodb]
django-celery==3.1.15
python-social-auth==0.2.1
django-rosetta==0.7.4
bokeh==0.6.1
numpy==1.9.1
panda==0.1.5
redis==2.10.3